+++ cogs/games/example.py
@@ -1,0 +1,9 @@
+"""Sub-cog example."""
+import gearbox
+cog = gearbox.Cog('example')
+
+
+@cog.command
+def example(author):
+    """Sub-cog command example."""
+    return f"Hi {author.name}, I'm a working example!"